Moving the annotations to common folder
