+++ maat-court-data-api/src/main/java/gov/uk/courtdata/applicant/controller/ApplicantController.java
@@ -1,6 +1,7 @@
 package gov.uk.courtdata.applicant.controller;
 
 import gov.uk.courtdata.annotation.NotFoundApiResponse;
+import gov.uk.courtdata.annotation.StandardApiResponseCodes;
 import gov.uk.courtdata.applicant.dto.ApplicantHistoryDTO;
 import gov.uk.courtdata.applicant.dto.RepOrderApplicantLinksDTO;
 import gov.uk.courtdata.applicant.dto.SendToCCLFDTO;
